pyproject.toml
[tool.poetry]
name = "order-info-page"
version = "0.1.0"
description = ""
authors = ["irumako <irumakots@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
pydantic-settings = "^2.2.1"
uvicorn = "^0.27.1"
jinja2 = "^3.1.3"
sqlalchemy = "^2.0.27"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pytest = "^8.0.2"
python-multipart = "^0.0.9"
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
